backend/requirements.txt
web3==7.14.0
eth-account==0.13.7
python-dotenv==1.2.1

